pyproject.toml
[tool.poetry]
name = "upl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^5.28.0"
Flask = "^2.1.1"
Flask-Cors = "^3.0.10"
gunicorn = "^20.1.0"
Flask-SocketIO = "^5.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
